Split DataUnionendPoints.test to stats.test and adminFee.test
